SchedulerFactory allows different scheduler implementations, other than the DefaultScheduler, to be used by the manager.
ManagerContainer, Configuration, and DefaultScheduler have been modified to accommodate or use the SchedulerFactory.

MappingScheduler provides a different scheduler implementation that maps applications to executors and subsequently schedules threads from the application to the mapped executors. Useful in scenarios where setting up an application to run on an executor is resource intensive.


git-svn-id: https://alchemi.svn.sourceforge.net/svnroot/alchemi/trunk@211 9c0b43c4-e927-0410-a2d5-cf972382494e